Updated firmware for wirefree DAQ
Added IR remote control on wirefree DAQ. Now the start and end of the recording can be triggered by two buttons on IR remote. Once the recording is ended by the IR control, it can not be restarted by the start button. You have to reconnect the power and press the start button